it's first commit. HibernateUtil has been created

+++ src/main/java/io/github/mat3e/HelloService.java
@@ -7,7 +7,7 @@ import java.util.logging.Logger;
 
 class HelloService {
     static final String FALLBACK_NAME = "world";
-    static final Lang FALLBACK_LANG = new Lang(1L, "Hello", "en");
+    static final Lang FALLBACK_LANG = new Lang(1, "Hello", "en");
     //private final Logger logger = (Logger) LoggerFactory.getLogger(HelloService.class);
 
     private LangRepository repository;
@@ -23,9 +23,9 @@ class HelloService {
     String prepareGreeting(String name, String lang) {
         //the parameter passed to the findById (that is what in the String lang - second parameter in the line above)
         // is of type Long so we have to convert  it from string to Long
-        Long langId;
+        Integer langId;
         try {
-            langId = Optional.ofNullable(lang).map(Long::valueOf).orElse(FALLBACK_LANG.getId()); //Long::valueOf returns passed String as Long
+            langId = Optional.ofNullable(lang).map(Integer::valueOf).orElse(FALLBACK_LANG.getId()); //Long::valueOf returns passed String as Long
         } catch(NumberFormatException e) {
             //logger.info("Non-numeric language is used" + lang);
             langId = FALLBACK_LANG.getId();
